refactor(im): move app.py IM init into cubebox/im/runtime.py
The 210-line _start_im_runtime / _stop_im_runtime closures lived in
api/app.py but were entirely about how IM is wired together: credential
cache, lark client cache, CardKit token plumbing, OutboundRunTailer
construction, FeishuLongConnection bootstrap. None of that concerns the
FastAPI factory.

Moves the implementation into cubebox/im/runtime.py with two entry
points (start / stop). app.py now imports cubebox.im.runtime at the call
sites and is back to being the FastAPI lifespan glue.

No behavior change.
